pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "agent-trade-sdk"
version = "0.1.0"
description = "Agentic paper-trading prototype with OpenAI Agents SDK, OpenRouter, Tavily and Alpaca."
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "openai-agents[litellm]>=0.9.2",
  "alpaca-py>=0.36.0",
  "tavily-python>=0.5.0",
  "python-dotenv>=1.0.1",
  "pydantic>=2.8.0",
]

[project.optional-dependencies]
dev = [
  "ruff>=0.8.0",
  "pytest>=8.0.0",
]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.ruff]
line-length = 100
target-version = "py311"
